+++ medusa/name_parser/rules/rules.py
@@ -397,6 +397,10 @@ class CreateAliasWithCountryOrYear(Rule):
         :type context: dict
         :return:
         """
+        # Don't add additional alias if we already have one from the previous rules
+        if matches.named('alias'):
+            return
+
         fileparts = matches.markers.named('path')
         for filepart in marker_sorted(fileparts, matches):
             title = matches.range(filepart.start, filepart.end, predicate=lambda match: match.name == 'title', index=0)
@@ -1083,12 +1087,19 @@ class FixParentFolderReplacingTitle(Rule):
             second_part = fileparts[parts_len - 2].value
             if self.ends_with_digit.search(second_part):
                 title = matches.named('title')
-                if not title or second_part.startswith(title[0].value):
+                if not title:
                     episode_title[0].name = 'title'
                     to_append = episode_title
-                    to_remove = title
-
+                    to_remove = None
                     return to_remove, to_append
+
+                if second_part.startswith(title[0].value):
+                    season = matches.named('season')
+                    if season and not second_part.endswith(season[-1].initiator.value):
+                        episode_title[0].name = 'title'
+                        to_append = episode_title
+                        to_remove = title
+                        return to_remove, to_append
 
 
 class FixMultipleSources(Rule):