+++ medusa/providers/torrent/rss/nyaa.py
@@ -32,7 +32,6 @@ class NyaaProvider(TorrentProvider):
 
         # Miscellaneous Options
         self.supports_absolute_numbering = True
-        self.anime_only = True
         self.confirmed = False
 
         # Torrent Stats
@@ -48,15 +47,19 @@ class NyaaProvider(TorrentProvider):
 
         :param search_strings: A dict with mode (key) and the search value (value)
         :param age: Not used
-        :param ep_obj: Not used
+        :param ep_obj: An episode object
         :returns: A list of search results (structure)
         """
         results = []
 
         # Search Params
+        category = '1_0'
+        if ep_obj and not ep_obj.series.is_anime:
+            category = '4_0'
+
         search_params = {
             'page': 'rss',
-            'c': '1_0',  # All Anime
+            'c': category,
             'f': 0,  # No filter
             'q': '',
         }