+++ medusa/server/web/home/add_shows.py
@@ -24,7 +24,6 @@ from medusa.show.recommendations.trakt import TraktPopular
 from medusa.show.show import Show
 
 from requests import RequestException
-from requests.compat import unquote_plus
 
 from simpleanidb import REQUEST_HOT
 
@@ -490,8 +489,12 @@ class HomeAddShows(Home):
         subtitles = config.checkbox_to_value(subtitles)
 
         if whitelist:
+            if not isinstance(whitelist, list):
+                whitelist = [whitelist]
             whitelist = short_group_names(whitelist)
         if blacklist:
+            if not isinstance(blacklist, list):
+                blacklist = [blacklist]
             blacklist = short_group_names(blacklist)
 
         if not allowed_qualities:
@@ -541,7 +544,7 @@ class HomeAddShows(Home):
         elif not isinstance(shows_to_add, list):
             shows_to_add = [shows_to_add]
 
-        shows_to_add = [unquote_plus(x) for x in shows_to_add]
+        shows_to_add = [text_type(x, 'utf-8') if not isinstance(x, text_type) else x for x in shows_to_add]
 
         prompt_for_settings = config.checkbox_to_value(prompt_for_settings)