+++ medusa/subtitles.py
@@ -563,18 +563,17 @@ def merge_subtitles(existing_subtitles, new_subtitles):
 def get_min_score():
     """Return the min score to be used by subliminal.
 
-    Perfect match = hash - resolution (subtitle for 720p is the same as for 1080p) - video_codec - audio_codec
+    Perfect match = series + year + season + episode + release_group
     Non-perfect match = series + year + season + episode
 
     :return: min score to be used to download subtitles
     :rtype: int
     """
+    min_score = episode_scores['series'] + episode_scores['year'] + episode_scores['season'] + episode_scores['episode']
     if app.SUBTITLES_PERFECT_MATCH:
-        return episode_scores['hash'] - (episode_scores['resolution'] +
-                                         episode_scores['video_codec'] +
-                                         episode_scores['audio_codec'])
+        min_score += episode_scores['release_group']
 
-    return episode_scores['series'] + episode_scores['year'] + episode_scores['season'] + episode_scores['episode']
+    return min_score
 
 
 def get_current_subtitles(tv_episode):